+++ src/examples/Program.cs
@@ -1,5 +1,9 @@
-﻿using PuppeteerReportCsharp;
+﻿using Konsole;
+using PuppeteerReportCsharp;
+using PuppeteerSharp;
+using System;
 using System.IO;
+using System.Net;
 using System.Reflection;
 using System.Threading.Tasks;
 
@@ -10,22 +14,42 @@ namespace examples
         static async Task Main(string[] args)
         {
             var example = "basicWithJs";
-            string filePath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), $"../../../../js-dependency/examples/{example}/index.html");
+            Console.WriteLine($"run example '{example}'");
 
+            string filePath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "..", "..", "..", "..", "js-dependency", "examples", example, "index.html");
+
+            var fether = new BrowserFetcher();
+
+            var pb = new ProgressBar(100);
+            void downloadProgressChanged(object sender, DownloadProgressChangedEventArgs e)
+            {
+                pb.Refresh(e.ProgressPercentage, "downloading Puppeteer");
+            }
+            fether.DownloadProgressChanged += downloadProgressChanged;
+
+            await fether.DownloadAsync();
+            pb.Refresh(100, "finished");
+
+            Console.WriteLine($"convert html to pdf");
             var puppeteer = new PuppeteerReport();
             await puppeteer.PDF(filePath,
-                "Result.pdf" , new PuppeteerSharp.PdfOptions
+                "Result.pdf", new PdfOptions
                 {
                     Format = PuppeteerSharp.Media.PaperFormat.A4,
                     PreferCSSPageSize = true,
                     MarginOptions = new PuppeteerSharp.Media.MarginOptions
                     {
-                        Top= "10mm",
+                        Top = "10mm",
                         Left = "10mm",
                         Right = "10mm",
                         Bottom = "10mm"
                     }
                 });
+
+            Console.WriteLine("finished");
+
+            string resultPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "Report.pdf");
+            Console.WriteLine(resultPath);
         }
     }
 }