+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.1] - 2021-07-04
+### Changed
+- update the base js project `puppeteer-report` to v 3.0.1
+- update `PuppeteerSharp` to 4.0.0
+- improve the examples project
+
+###Removed
+- `BrowserFetcher` has been removed. you need to fetch the puppeteer before calling the pdf method.
+
 ## [1.1.1] - 2020-10-09
 ### Added
 - Add a new `pdfPage` method to use custom puppeteer page instance